separate CI classes
