+++ src/main/treeDataProviders/treesManager.ts
@@ -3,9 +3,9 @@ import { ConnectionManager } from '../connect/connectionManager';
 import { ExtensionComponent } from '../extensionComponent';
 import { LogManager } from '../log/logManager';
 import { ScanCacheManager } from '../cache/scanCacheManager';
-import { BuildsDataProvider } from './dependenciesTree/buildsDataProvider';
+import { BuildsDataProvider } from './ciTree/buildsDataProvider';
 import { DependenciesTreeNode } from './dependenciesTree/dependenciesTreeNode';
-import { DependencyDetailsProvider } from './dependencyDetailsProvider';
+import { DependencyDetailsProvider } from './ciTree/dependencyDetailsProvider';
 import { IssuesTreeDataProvider } from './issuesTree/issuesTreeDataProvider';
 import { FileTreeNode } from './issuesTree/fileTreeNode';
 import { ScanManager } from '../scanLogic/scanManager';
