+++ src/main/utils/consts.ts
@@ -1,4 +1,4 @@
-export class Consts {
+export class Constants {
     public static readonly MISSING_COMPONENT: string = 'Component is missing in Xray';
     public static readonly UNKNOWN_LICENSE_FULL_NAME: string = 'Unknown license';
     public static readonly UNKNOWN_LICENSE: string = 'Unknown';
